refactor(proxy): extract telemetry types to metrics.go
- Move tunnelEntry struct to metrics.go
- Move countingWriter struct and Write method to metrics.go
- Move classifyProxyError function to metrics.go
- Add flush() method to tunnelEntry for cleaner byte flushing
- Add buildAttrs() method to tunnelEntry for attribute initialization
- Update countingWriter to use asyncBytes and entry fields
- Simplify flushLoop and ClearTunnelID to use flush() method
